+++ drivers/adodb-sqlite3.inc.php
@@ -1,6 +1,8 @@
 <?php
 /*
-V5.20dev  ??-???-2014  (c) 2000-2014 John Lim (jlim#natsoft.com). All rights reserved.
+@version   v5.21dev  ??-???-2015
+@copyright (c) 2000-2013 John Lim (jlim#natsoft.com). All rights reserved.
+@copyright (c) 2014      Damien Regad, Mark Newnham and the ADOdb community
   Released under both BSD license and Lesser GPL library license.
   Whenever there is any discrepancy between the two licenses,
   the BSD license will take precedence.
@@ -363,7 +365,6 @@ class ADORecordset_sqlite3 extends ADORecordSet {
 
 	protected function _initrs()
 	{
-		$this->_numOfRows = 1;
 		$this->_numOfFields = $this->_queryID->numColumns();
 
 	}